+++ configuration/src/main/java/com/synopsys/integration/configuration/config/PropertyConfiguration.java
@@ -192,7 +192,16 @@ public class PropertyConfiguration {
         Map<String, String> rawMap = new HashMap<>();
         for (Property property : properties) {
             String rawKey = property.getKey();
-            getRaw(property).ifPresent(rawValue -> rawMap.put(rawKey, maskValue(rawKey, rawValue, shouldMask)));
+            if (property instanceof PassthroughProperty) {
+                getRaw((PassthroughProperty) property).entrySet()
+                    .forEach(passThrough -> {
+                            String fullPassThroughKey = String.format("%s.%s", rawKey, passThrough.getKey());
+                            rawMap.put(fullPassThroughKey, maskValue(fullPassThroughKey, passThrough.getValue(), shouldMask));
+                        }
+                    );
+            } else {
+                getRaw(property).ifPresent(rawValue -> rawMap.put(rawKey, maskValue(rawKey, rawValue, shouldMask)));
+            }
         }
         return rawMap;
     }