+++ detectable/src/main/java/com/synopsys/integration/detectable/detectables/dart/pubspec/PubSpecExtractor.java
@@ -1,11 +1,13 @@
 package com.synopsys.integration.detectable.detectables.dart.pubspec;
 
 import java.io.File;
+import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.util.List;
 import java.util.Optional;
 
+import org.jetbrains.annotations.Nullable;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -18,27 +20,26 @@ import com.synopsys.integration.util.NameVersion;
 public class PubSpecExtractor {
     private final Logger logger = LoggerFactory.getLogger(this.getClass());
 
-    private PubSpecLockParser pubSpecLockParser;
-    private PubSpecYamlNameVersionParser nameVersionParser;
+    private final PubSpecLockParser pubSpecLockParser;
+    private final PubSpecYamlNameVersionParser nameVersionParser;
 
     public PubSpecExtractor(PubSpecLockParser pubSpecLockParser, PubSpecYamlNameVersionParser nameVersionParser) {
         this.pubSpecLockParser = pubSpecLockParser;
         this.nameVersionParser = nameVersionParser;
     }
 
-    public Extraction extract(File pubSpecLockFile, File pubSpecYamlFile) {
-        try {
-            List<String> pubSpecLockLines = Files.readAllLines(pubSpecLockFile.toPath(), StandardCharsets.UTF_8);
-            logger.debug(String.join(System.lineSeparator(), pubSpecLockLines));
+    public Extraction extract(File pubSpecLockFile, @Nullable File pubSpecYamlFile) throws IOException {
+        List<String> pubSpecLockLines = Files.readAllLines(pubSpecLockFile.toPath(), StandardCharsets.UTF_8);
 
-            Optional<NameVersion> nameVersion = nameVersionParser.parseNameVersion(pubSpecYamlFile);
-
-            DependencyGraph dependencyGraph = pubSpecLockParser.parse(pubSpecLockLines);
-
-            CodeLocation codeLocation = new CodeLocation(dependencyGraph);
-            return new Extraction.Builder().success(codeLocation).nameVersionIfPresent(nameVersion).build();
-        } catch (Exception e) {
-            return new Extraction.Builder().exception(e).build();
+        Optional<NameVersion> nameVersion = Optional.empty();
+        if (pubSpecYamlFile != null) {
+            List<String> pubSpecYamlLines = Files.readAllLines(pubSpecYamlFile.toPath(), StandardCharsets.UTF_8);
+            nameVersion = nameVersionParser.parseNameVersion(pubSpecYamlLines);
         }
+
+        DependencyGraph dependencyGraph = pubSpecLockParser.parse(pubSpecLockLines);
+
+        CodeLocation codeLocation = new CodeLocation(dependencyGraph);
+        return new Extraction.Builder().success(codeLocation).nameVersionIfPresent(nameVersion).build();
     }
 }