+++ src/main/java/com/synopsys/integration/detect/tool/signaturescanner/operation/PublishSignatureScanReports.java
@@ -20,10 +20,12 @@ public class PublishSignatureScanReports {
     private final Logger logger = LoggerFactory.getLogger(this.getClass());
     private final ExitCodePublisher exitCodePublisher;
     private final StatusEventPublisher statusEventPublisher;
+    private final boolean treatSkippedScanAsSuccess;
 
-    public PublishSignatureScanReports(ExitCodePublisher exitCodePublisher, StatusEventPublisher statusEventPublisher) {
+    public PublishSignatureScanReports(ExitCodePublisher exitCodePublisher, StatusEventPublisher statusEventPublisher, boolean treatSkippedScanAsSuccess) {
         this.exitCodePublisher = exitCodePublisher;
         this.statusEventPublisher = statusEventPublisher;
+        this.treatSkippedScanAsSuccess = treatSkippedScanAsSuccess;
     }
 
     public void publishReports(List<SignatureScannerReport> signatureScannerReports) {
@@ -39,16 +41,33 @@ public class PublishSignatureScanReports {
                 ));
                 exitCodePublisher.publishExitCode(new ExitCodeRequest(ExitCodeType.FAILURE_SCAN));
             });
+
+        if (!treatSkippedScanAsSuccess) {
+            signatureScannerReports.stream()
+                .filter(SignatureScannerReport::isSkipped)
+                .findAny()
+                .ifPresent(report -> {
+                    logger.error(
+                        "The Signature Scanner skipped a scan because the minimum scan interval was not met.");
+                    exitCodePublisher.publishExitCode(new ExitCodeRequest(ExitCodeType.FAILURE_MINIMUM_INTERVAL_NOT_MET));
+                });
+        }
     }
 
     private void publishReport(SignatureScannerReport signatureScannerReport) {
-        if (signatureScannerReport.isSuccessful()) {
+        if (signatureScannerReport.isSuccessful() || (signatureScannerReport.isSkipped() && treatSkippedScanAsSuccess)) {
             statusEventPublisher.publishStatusSummary(new SignatureScanStatus(signatureScannerReport.getSignatureScanPath().getTargetCanonicalPath(), StatusType.SUCCESS));
             return;
         }
 
         String scanTargetPath = signatureScannerReport.getSignatureScanPath().getTargetCanonicalPath();
-        if (!signatureScannerReport.hasOutput()) {
+        if (signatureScannerReport.isSkipped()) {
+            statusEventPublisher.publishIssue(new DetectIssue(
+                DetectIssueType.SIGNATURE_SCANNER,
+                String.format("Scanning target %s was never scanned by the BlackDuck CLI.", scanTargetPath),
+                Arrays.asList("The minimum scan interval was not met and this scan was skipped by the BlackDuck CLI.")
+            ));
+        } else if (!signatureScannerReport.hasOutput()) {
             String errorMessage = String.format("Scanning target %s was never scanned by the BlackDuck CLI.", scanTargetPath);
             logger.info(errorMessage);
             statusEventPublisher.publishIssue(new DetectIssue(DetectIssueType.SIGNATURE_SCANNER, "Black Duck Signature Scanner", Arrays.asList(errorMessage)));