+++ src/main/java/com/synopsys/integration/detect/tool/signaturescanner/operation/SignatureScanOuputResult.java
@@ -1,21 +1,12 @@
 package com.synopsys.integration.detect.tool.signaturescanner.operation;
 
-import java.util.List;
-
 import com.synopsys.integration.blackduck.codelocation.signaturescanner.ScanBatchOutput;
-import com.synopsys.integration.blackduck.codelocation.signaturescanner.command.ScanCommandOutput;
 
 public class SignatureScanOuputResult {
-    private final List<ScanCommandOutput> scanCommandOutputs;
     private final ScanBatchOutput scanBatchOutput;
 
-    public SignatureScanOuputResult(List<ScanCommandOutput> scanCommandOutputs, ScanBatchOutput scanBatchOutput) {
-        this.scanCommandOutputs = scanCommandOutputs;
+    public SignatureScanOuputResult(ScanBatchOutput scanBatchOutput) {
         this.scanBatchOutput = scanBatchOutput;
-    }
-
-    public List<ScanCommandOutput> getScanCommandOutputs() {
-        return scanCommandOutputs;
     }
 
     public ScanBatchOutput getScanBatchOutput() {