+++ docs/markdown/releasenotes.md
@@ -5,12 +5,20 @@
 ### New features
 
 * Added support for a buildless Pipenv detector that parses the Pipfile.lock file (see [python support page](packagemgrs/python.md) for more details).
+* [solution_name] now includes pass-through properties when logging configuration at the beginning of a run.
 
 ### Changed features
 
 ### Resolved issues
 
 * (IDETECT-3184) Resolved an issue that prevented matches for Bazel maven_install components with complex (>3 parts) maven_coordinates values.
+* (IDETECT-3207) Resolved an issue that prevented Bazel and Docker Tool issues from being reported in the issues section of the [solution_name] log and status file.
+
+## Version 7.12.1
+
+## Changed Features
+
+* When signature scanning is skipped due to the minimum scan interval on Black Duck not being met, [solution_name] by default will treat the run as a success and will not wait for the skipped scan(s). If the property detect.force.success.on.skip is set to false, [solution_name] will instead return exit code 13 when one or more signature scans were skipped.
 
 ## Version 7.12.0